Move conversion of AST to JSON into its class, add command line option
